services/auth-service/pyproject.toml
[project]
name = "auth-service"
version = "0.1.0"
description = ""
authors = [
    {name = "Gustavo Emanuel Silva Queiroz", email = "gustavoe.dev@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
fastapi = ">=0.121.2,<0.122.0"
uvicorn = ">=0.38.0,<0.39.0"
pydantic-settings = ">=2.12.0,<3.0.0"
alembic = ">=1.13.0"
psycopg2-binary = ">=2.9.0"
asyncpg = ">=0.29.0"
python-jose = "==3.5.0"
python-keycloak = "==5.8.1"
email-validator = "^2.1.0"
python-slugify = "==8.0.4"
database-lib = {path = "../../libs/database", develop = true}
common-lib = {path = "../../libs/common", develop = true}
resend = "^2.19.0"
jinja2 = "^3.1.6"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.24.0"
pytest-cov = "^6.0.0"
httpx = "^0.28.0"
aiosqlite = "^0.20.0"

[tool.coverage.run]
source = ["src"]
branch = true
omit = [
    "src/config/*",
    "src/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
show_missing = true

[tool.coverage.html]
directory = "htmlcov"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
